% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-highcharter.R
\name{hc_stern_cat}
\alias{hc_stern_cat}
\title{Apply Stern categorical palette to a Highchart object.}
\usage{
hc_stern_cat(hc)
}
\arguments{
\item{hc}{A Highchart object.}
}
\value{
The Highchart object, with palette applied.
}
\description{
Sets the \code{colors} option to the categorical scale in use order. Use
when you want explicit control independent of \code{\link[=hc_theme_stern]{hc_theme_stern()}}.
}
\details{
Requires the optional package \code{highcharter}.
}
