% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-highcharter.R
\name{hc_stern_div}
\alias{hc_stern_div}
\title{Apply Stern diverging palette to a Highchart object.}
\usage{
hc_stern_div(hc, reverse = FALSE)
}
\arguments{
\item{hc}{A Highchart object.}

\item{reverse}{Logical. If \code{TRUE}, navy maps to first series and olive
to last. Default \code{FALSE}.}
}
\value{
The Highchart object, with palette applied.
}
\description{
For series where order spans a meaningful midpoint.
}
\details{
Requires the optional package \code{highcharter}.
}
