% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-highcharter.R
\name{hc_stern_seq_stops}
\alias{hc_stern_seq_stops}
\title{Sequential color stops for \code{hc_colorAxis()}.}
\usage{
hc_stern_seq_stops()
}
\value{
A list of \verb{(position, color)} pairs.
}
\description{
Use when you have a continuous fill (heatmap, treemap, choropleth) and
need to pass \code{stops} to \code{highcharter::hc_colorAxis()}.
}
\examples{
\dontrun{
library(highcharter)

hchart(my_data, "heatmap", hcaes(x = x, y = y, value = z)) |>
  hc_colorAxis(stops = hc_stern_seq_stops()) |>
  hc_add_theme(hc_theme_stern())
}

}
