% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{stern_border}
\alias{stern_border}
\title{Standard border color.}
\format{
A character scalar -- \code{"#C9C0A4"}.
}
\usage{
stern_border
}
\description{
Panel borders, axis lines, card and table borders.
}
