% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-shiny.R
\name{stern_custom_css}
\alias{stern_custom_css}
\title{Custom CSS rules layered on top of the bs_theme.}
\usage{
stern_custom_css(base_size = "16px")
}
\arguments{
\item{base_size}{CSS base font size. Default \code{"16px"}.}
}
\value{
A character scalar of CSS.
}
\description{
Returns a single string of CSS. Used internally by \code{\link[=stern_bs_theme]{stern_bs_theme()}},
but exposed in case you want to inject the rules into a non-bslib
context (e.g. an \code{htmltools::tagList()} with \code{tags$style()}).
}
