% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ggplot.R
\name{theme_stern_map}
\alias{theme_stern_map}
\title{Stern theme variant -- for maps and spatial visualizations.}
\usage{
theme_stern_map(base_size = 14, legend = "bottom")
}
\arguments{
\item{base_size}{Numeric. Base font size in points. Default 12.}

\item{legend}{Character. Legend position. See \code{\link[=theme_stern]{theme_stern()}}.}
}
\value{
A \code{ggplot2::theme} object.
}
\description{
Same as \code{\link[=theme_stern]{theme_stern()}} but with no gridlines, no panel border, and no
axis ticks or labels. Use for choropleths and other spatial plots.
}
\examples{
\dontrun{
library(ggplot2)
library(sf)
stern_setup_fonts()

ggplot(my_sf_data) +
  geom_sf(aes(fill = value), color = stern_border) +
  scale_fill_stern_seq() +
  theme_stern_map()
}

}
